// src/enums/game.ts
export enum LiveStreamEnum {
  RikVgmn108 = 'rik_vgmn_108',
  RikVgmn109 = 'rik_vgmn_109',
  RikVgmn110 = 'rik_vgmn_110',
  RikVgmn111 = 'rik_vgmn_111',
  GoQsTxgo101 = 'go_qs_txgo-101',
  GoVgmn109 = 'go_vgmn_109',
  B52Vgmn108 = 'b52_vgmn_108',
  B52Vgmn109 = 'b52_vgmn_109',
  B52Vgmn110 = 'b52_vgmn_110',
  Game789G1X305 = '789game_G1X_305',
  Game789G1X306 = '789game_G1X_306',
  Vgmn108 = 'vgmn_108',
  Vgmn109 = 'vgmn_109',
  Vgmn110 = 'vgmn_110',
  Vgmn111 = 'vgmn_111',
  QsTxgo101 = 'qs_txgo-101',
  QsXocdia102 = 'qs_xocdia-102',
  G1X305 = 'G1X_305',
  G1X306 = 'G1X_306',
  G1S305 = 'G1S_305',
  G1S306 = 'G1S_306',
  G1S311 = 'G1S_311',
  SungameG1S311 = 'sungame_G1S_311',
  SungameG1S305 = 'sungame_G1S_305',
  SungameG1S306 = 'sungame_G1S_306',
}